make more or less portable

+++ pt/scripts/monitor_exclusivity.pl
@@ -2,30 +2,56 @@
 
 use strict;
 
-use lib "$ENV{SDRROOT}/mdp-lib/Utils";
-use Vendors;
-use Database;
-
 # use Try::Tiny;
 use IO::File;
 use IPC::Run qw(run);
-use Date::Manip;
+use POSIX qw(strftime);
 use Mail::Mailer;
+use Data::Dumper;
+use DBI;
+
+use Getopt::Long;
 
 my $TABLE_NAME = q{pt_exclusivity_ng};
-my $LOGDIR = "$ENV{SDRROOT}/logs/pt_exclusivity";
+my $LOGDIR = "/htprep/stats/pt_exclusivity";
+my $is_reporting = 0;
+my $is_debugging = 0;
+
+GetOptions(
+    "logdir=s" => \$LOGDIR,
+    "report" => \$is_reporting,
+    "debug" => \$is_debugging,
+);
+
 if ( ! -d $LOGDIR ) {
-    mkdir($LOGDIR, 0775);
+    exit;
 }
 my $TMPFILE = "$LOGDIR/report.tmp";
 
 my @now =localtime(time);
-my $is_reporting = ( ( $now[2] == 23 && $now[1] >= 45 ) || $ARGV[0] eq '--report' );
+$is_reporting = ( $now[2] == 23 && $now[1] >= 45 ) unless ( $is_reporting );
+
+my $config = {};
+open(my $in, "<", "/htapps/babel/etc/ht_web.conf") or die "could not open ht_web.conf - $!";
+while ( my $line = <$in> ) {
+  chomp $line;
+  next if ( $line =~ m,^#, || $line =~ m,^$, );
+  my ( $key, $value ) = split(/\s+=\s+/, $line);
+  $$config{$key} = $value;
+}
+close($in);
 
 # Database connection
-my $db =  new Database('ht_web');
-my $dbh = $db->get_DBH();
-$dbh->{RaiseError} = 1;
+my $dsn = qq{DBI:mysql:$$config{db_name}:$$config{db_server}};
+my $dbh = DBI->connect(
+                    $dsn,
+                    $$config{db_user},
+                    $$config{db_passwd},
+                  {
+                   RaiseError => 1,
+                   PrintError => 0,
+                  }
+                   );
 
 my $monitor_sql = <<SQL;
 SELECT lock_id, item_id, owner, affiliation, expires, renewals FROM $TABLE_NAME ORDER BY affiliation, expires, owner, renewals;
@@ -43,11 +69,17 @@ foreach my $row ( @$rows ) {
 $fh->close;
 
 if ( $is_reporting ) {
-    my $timestamp = UnixDate("now", "%Y-%m-%d");
+    # my $timestamp = UnixDate("now", "%Y-%m-%d");
+    my $timestamp = strftime("%Y-%m-%d", @now);
     my $filename = "$LOGDIR/grants-$timestamp.log";
     my $err;
     my $body;
-    run [ "/usr/bin/sort", "-u", $TMPFILE ], \undef, \$body, \$err;
+
+    # only include today's data
+    run 
+      [ "/usr/bin/sort", "-u", $TMPFILE ], "|", [ "egrep", $timestamp ],
+      \$body, \$err;
+
     if ( $body ) {
         # mail this file out
         my $mailer = new Mail::Mailer('sendmail');
@@ -70,7 +102,7 @@ if ( $is_reporting ) {
         print $fh $body, "\n";
         $fh->close;
     }
-    unlink($TMPFILE);
+    unlink($TMPFILE) unless ( $is_debugging );
 }
